iOS7にして出たwarning対応

+++ SMCalloutView.m
@@ -543,15 +543,15 @@ NSTimeInterval kSMCalloutViewRepositionDelayForUIScrollView = 1.0/3.0;
     };
     
     NSData *inputData = [string dataUsingEncoding:NSASCIIStringEncoding allowLossyConversion:YES];
-    long long inputLength = [inputData length];
+    NSInteger inputLength = [inputData length];
     const unsigned char *inputBytes = [inputData bytes];
     
-    long long maxOutputLength = (inputLength / 4 + 1) * 3;
+    NSInteger maxOutputLength = (inputLength / 4 + 1) * 3;
     NSMutableData *outputData = [NSMutableData dataWithLength:maxOutputLength];
     unsigned char *outputBytes = (unsigned char *)[outputData mutableBytes];
     
     int accumulator = 0;
-    long long outputLength = 0;
+    NSInteger outputLength = 0;
     unsigned char accumulated[] = {0, 0, 0, 0};
     for (long long i = 0; i < inputLength; i++) {
         unsigned char decoded = lookup[inputBytes[i] & 0x7F];
